+++ script/publish.clj
@@ -22,8 +22,8 @@
       (nth 0)
       str/trim))
 
-(defn update-changelog
-  [file-name changelog-text unreleased-header-re version]
+(defn new-changelog-text
+  [changelog-text unreleased-header-re version]
   (println "Updating changelog")
   (let [utc-date (-> (java.time.Instant/now)
                      .toString
@@ -33,8 +33,8 @@
         new-text (str/replace-first
                   changelog-text
                   unreleased-header-re
-                  (format "[Unreleased]\n\n%s\n" new-header))]
-    (spit file-name new-text)))
+                  (format "[Unreleased]\n\n%s\n\n" new-header))]
+    new-text))
 
 (defn throw-if-error [{:keys [exit out err]}]
   (when-not (= exit 0)
@@ -65,22 +65,22 @@
   (println "Open to follow the progress of the release:")
   (println "  https://app.circleci.com/pipelines/github/BetterThanTomorrow/joyride"))
 
-(let [unreleased-changelog-text (get-unreleased-changelog-text
-                                 changelog-text
-                                 unreleased-header-re)]
-  (if (empty? unreleased-changelog-text)
-    (do
-      (print "There are no unreleased changes in the changelog. Release anyway? y/n: ")
-      (flush)
-      (let [answer (read)]
-        (if (= (str answer) "y")
-          (publish)
-          (println "Aborting publish."))))
-    (do
-      (update-changelog changelog-filename
-                        changelog-text
-                        unreleased-header-re
-                        joyride-version)
-      (commit-changelog changelog-filename
-                        (str "Add changelog section for v" joyride-version " [skip ci]"))
-      (publish))))
+(when (= *file* (System/getProperty "babashka.file"))
+  (let [unreleased-changelog-text (get-unreleased-changelog-text
+                                   changelog-text
+                                   unreleased-header-re)]
+    (if (empty? unreleased-changelog-text)
+      (do
+        (print "There are no unreleased changes in the changelog. Release anyway? y/n: ")
+        (flush)
+        (let [answer (read)]
+          (if (= (str answer) "y")
+            (publish)
+            (println "Aborting publish."))))
+      (let [updated-changelog-text (new-changelog-text changelog-text
+                                                       unreleased-header-re
+                                                       joyride-version)]
+        (spit changelog-filename updated-changelog-text)
+        (commit-changelog changelog-filename
+                          (str "Add changelog section for v" joyride-version " [skip ci]"))
+        (publish)))))