+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes to Joyride
 - [Add Workspace script menu items for creating getting started content](https://github.com/BetterThanTomorrow/joyride/issues/57)
 
 ## [0.0.9] - 2022-05-11
+
 - [Error when dismissing the scripts menu without selecting anything](https://github.com/BetterThanTomorrow/joyride/issues/24)
 - Fix: [Better error report when a script isn't found](https://github.com/BetterThanTomorrow/joyride/issues/40)
 - [nrepl: support pprinting eval results](https://github.com/BetterThanTomorrow/joyride/issues/49)